+++ ChangeLog.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.38.0]
 ### Changed
 - `pyvicp` is now an optional dependency in order to maintain consistency with available conda packages
 - doc updates and fixes
+- minor bugfixes to `repr()` 
+- corrected an exception for some `labbench.paramattr.method` decorator implementations
+- `AttributeError` is now raised when assigning to `method` paramattrs as if they were `property` paramattrs
 
 ## [0.37.0]
 ### Changed